refactor(country-pick): drop dead URL branch from pickUnvisited
AC 7 moved URL validation into ChartScreen (validateUrlCode), so the
random-pick layer below only ever saw urlParam=null. resolveCountry
becomes pickUnvisited — same algorithm, narrower contract (pick from
allCodes minus visited, full reset fallback when exhausted). The
ResolveSource type and source field are gone; only "random" existed.

File renamed country-resolution.ts → pick-unvisited.ts.
